pyproject.toml
[tool.poetry]
name = "python-modeling"
version = "0.1.0"
description = ""
authors = ["Maxim Mozgovoy <mozgovoy@u-aizu.ac.jp>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.9"

[tool.poetry.dev-dependencies]
plumbum = "^1.8.2"
PyYAML = "^6.0.1"
pandoc-minted = "^0.2.0"
pandoc-crossref = "^0.1.1"

[tool.poetry.group.dev.dependencies]
pygments = "^2.17.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
